requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
Pillow>=9.5.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
whitenoise>=6.4.0
django-filter>=23.0.0
markdown>=3.4.0
dj-database-url>=2.0.0
drf-yasg>=1.21.0
